Puppeteer and PageSpeed Data Collection
Squashed commit with the results of work done on the dev branch.
Contains the code to collect data using Puppeteer and the
GooglePageSpeed API.

Next commit will contain the url dataset which will be used to take
data.

Results are written to files in both raw and aggregated format. Each set
of results is stored in its own directory.

Added elements for mobile emulation also. Results are stored in separate
folders to differentiate easily between them.