Big commit for Puppeteer and Pagespeed
Significant progress for taking data from both Puppeteer and the PageSpeed
API. Produces raw and aggregated data for the urls given in the urls.txt
file. Also performs each component several times to get an average of
the metrics produces from Puppeteer and PageSpeed.

Added the experimentProcess and experimentResults interfaces to try and
standardise things. More work might be done in future on this.

Added elements for mobile emulation also. Results are stored in separate
folders to differentiate easily between them.