URL Dataset and Cleanup
Added the URL dataset of the top 500 websites in Ireland. To make them
work with Puppeteer I've updated the preprocessing to prepend http:// to
the url. This enables redirection to https://

Also added a filter for google websites, i.e. google.com, google.ie etc.
This is due to an issue experience with the PageSpeed API that I've only
seen happen with Google sites. Will investigate further at some point.

Modified the main loop to run the each experiment asynchronously, this
is to allow both Puppeteer and PageSpeed to run simulataneously.