Puppeteer Cleanup + Documentation
Modified how the Puppeteer component is started. It now allows for a
comma separated list of URLS to be passed to the program. In this case
the experiment runs for each of those URLS. In this scenario the
completed and failed lists have no impact on what URLS get tested.

Another option is to pass the path of a file containing the URLS to run
the experiment with. In this mode, the experiment utilises the completed
and failed lists for handling issues that may arise during the
experiment. If the experiment fails, then on the next startup it will
continue from after the last url ran in the previous run.

The path to the browser must now be passed. The constant will be removed
in a future commit.

The wipeData argument results in the completed and failed list to be
wiped before starting the experiment.

Updated the InitializeDirectories function to be a bit more useful.